refactor: extract frontmatter lift helpers
